Properly structure directory and fix paths for deployment
