+++ city-transit/src/components/bus-route-map/BusRouteMap.tsx
@@ -3,20 +3,15 @@ import { useState } from "react";
 import { transitRoutes } from "../../data/transitData";
 
 export default function BusRouteMap() {
-
-  // currently selected route
   const [selectedRouteId, setSelectedRouteId] = useState("BLUE");
 
-  // find route object
   const selectedRoute = transitRoutes.find(r => r.id === selectedRouteId)!;
 
-  // convert stops → polyline positions
   const positions = selectedRoute.stops.map(stop =>
     [stop.lat, stop.lng] as [number, number]
   );
 
   return (
-
     <section>
       <h2>Bus Route Map</h2>
 
@@ -35,7 +30,7 @@ export default function BusRouteMap() {
       {/* Map */}
       <div style={{ height: "500px", width: "100%", marginTop: "10px" }}>
         <MapContainer
-          key={selectedRoute.id}   // forces map refresh when route changes
+          key={selectedRoute.id}
           center={positions[0]}
           zoom={12}
           style={{ height: "100%", width: "100%" }}
@@ -50,42 +45,6 @@ export default function BusRouteMap() {
             <Marker key={`${selectedRoute.id}-${stop.id}`} position={[stop.lat, stop.lng]}>
               <Popup>{stop.name}</Popup>
             </Marker>
-
-    <section className="bus-route-map" aria-labelledby="route-map-title">
-      <h2 id="route-map-title">Bus Route Map</h2>
-
-      <div className="map-area">
-        <svg
-          width="100"
-          height="180"
-          role="img"
-          aria-label="Static illustration of bus routes"
-        >
-          <line
-            x1="50"
-            y1="20"
-            x2="50"
-            y2="160"
-            stroke="#003A8F"
-            strokeWidth="4"
-          />
-          {stops.map((stop, index) => (
-            <circle
-              key={stop.id}
-              cx="50"
-              cy={20 + index * 35}
-              r="6"
-              fill="#003A8F"
-            />
-          ))}
-        </svg>
-
-        <ul className="stop-list">
-          {stops.map((stop, index) => (
-            <li key={stop.id} style={{ marginTop: index === 0 ? '0' : '1rem' }}>
-              <button type="button">{stop.name}</button>
-            </li>
-
           ))}
 
           {/* Route Line */}
@@ -94,7 +53,6 @@ export default function BusRouteMap() {
             positions={positions}
             pathOptions={{ color: selectedRoute.color, weight: 5 }}
           />
-
         </MapContainer>
       </div>
     </section>
