--- city-transit/src/components/bus-route-map/live-bus-tracker/nav/nav.tsx
@@ -1,3 +1,0 @@
-export default function Nav() {
-  return null;
-}
