+++ city-transit/src/repositories/tripRepository.ts
